Reorder client escape menu
Remove the unfinished resync action, place Steam invites below Return to the Game, and keep save disabling independent of optional menu entries.\n\nRefs #1857
