+++ source/GameInterface/Services/UI/Patches/EscapeMenuCoopOptionsPatch.cs
@@ -1,4 +1,4 @@
-using Common;
+﻿using Common;
 using GameInterface.Services.UI.CoopOptions;
 using HarmonyLib;
 using SandBox.View.Map;
@@ -18,19 +18,24 @@ namespace GameInterface.Services.UI.Patches;
 [HarmonyPatch(typeof(MapScreen), "GetEscapeMenuItems")]
 internal class EscapeMenuCoopOptionsPatch
 {
+    private static readonly TextObject CampaignOptionsText = new TextObject("{=PXT6aA4J}Campaign Options");
+
     [HarmonyPostfix]
     static void AddCoopOptionsItem(List<EscapeMenuItemVM> __result)
     {
         if (ModInformation.IsServer) return;
 
-        __result.Insert(2, new EscapeMenuItemVM(
+        int campaignOptionsIndex = __result.FindIndex(item => item.ActionText == CampaignOptionsText.ToString());
+        if (campaignOptionsIndex < 0) return;
+
+        __result.Insert(campaignOptionsIndex + 1, new EscapeMenuItemVM(
             new TextObject("Coop Options"),
             _ => ScreenManager.PushScreen(ViewCreatorManager.CreateScreenView<CoopOptionsUI>()),
             identifier: null,
             getIsDisabledAndReason: () => new Tuple<bool, TextObject>(false, new TextObject("")),
             isPositiveBehaviored: false));
 
-        EscapeMenuResyncButtonPatch.CustomButtonInserted = true;
+        EscapeMenuPanelHeightPatch.CustomButtonInserted = true;
     }
 }
