+++ source/GameInterface/Services/UI/Patches/EscapeMenuDisableSavePatch.cs
@@ -1,4 +1,4 @@
-using Common;
+﻿using Common;
 using HarmonyLib;
 using SandBox.View.Map;
 using System;
@@ -14,12 +14,11 @@ namespace GameInterface.Services.UI.Patches
     [HarmonyPatch(typeof(MapScreen), "GetEscapeMenuItems")]
     internal class EscapeMenuDisableSavePatch
     {
-        private const int SaveButtonIndex = 4;
-        private const int SaveAsButtonIndex = 5;
-
         // EscapeMenuItemVM.RefreshValues() recomputes IsDisabled from the item's disabled-func
         // on every refresh, so simply setting IsDisabled = true would be overwritten. Instead
         // the item is replaced with a clone whose func always reports disabled.
+        private static readonly TextObject SaveButtonText = new TextObject("{=bV75iwKa}Save");
+        private static readonly TextObject SaveAsButtonText = new TextObject("{=e0KdfaNe}Save As");
         private static readonly TextObject DisabledReason =
             new TextObject("Saving is disabled on clients; the host saves the campaign.");
 
@@ -28,13 +27,16 @@ namespace GameInterface.Services.UI.Patches
         {
             if (!ModInformation.IsClient) return;
 
-            DisableItem(__result, SaveButtonIndex);
-            DisableItem(__result, SaveAsButtonIndex);
+            DisableItem(__result, SaveButtonText);
+            DisableItem(__result, SaveAsButtonText);
         }
 
-        private static void DisableItem(List<EscapeMenuItemVM> items, int index)
+        private static void DisableItem(List<EscapeMenuItemVM> items, TextObject buttonText)
         {
-            if (index < 0 || index >= items.Count) return;
+            // Invite Friends is optional and inserted before these entries, so their indices vary.
+            string actionText = buttonText.ToString();
+            int index = items.FindIndex(item => item.ActionText == actionText);
+            if (index < 0) return;
 
             EscapeMenuItemVM original = items[index];
             items[index] = new EscapeMenuItemVM(
